requirements.txt
# Function dependencies, for example:
# package>=version

google-cloud-datacatalog
google-cloud-core
google-cloud-bigquery
google-cloud-storage
gcsfs
pandas
